refactor: util 클래스는 common/util 패키지로 옮기기
